Expand the test module
Expand the test module to allow the creation of test suites and
tests that group together multiple checks. While the check and
checkRaises functions work well for less complex code, some of
the more complex classes in the package (such as Grammar, which
essentially implements a DSL) require more heavy duty tests.