Add ConstantRanker and fix a related bug in the demo
The demo was broken because it expected the parsers to be wrapped
in rankers but the digitParser and ignorableParser didn't have rankers
because they only return one parse. To fix this issue, ConstantRanker
ranks all parses equally and provides an easy way to wrap a parser in
the ranker interface when no real ranking logic is necessary.