Improve the Ranker interface and update existing rankers
Because Rankers are trained to discriminate the correct parse among
the set of parses produced from one parser for one utterance, the
notion of scoring a parse independent of it's context (i.e., the other
parses produced by the parser for that utterance) isn't really
meaningful. To prevent confusion, it's better not to make the score
function part of the public interface, instead leaving it as an
implementation detail. Correspondingly, the score functions for
existing rankers have been changed to private functions.

Moreoever, depending on the model a ranker implements, there are
different ways of ranking denotations after generating scores for the
parses. For example, the SoftmaxRanker is trained via maximum marginal
likelihood on the denotations, so denotations should be ranked by
marginalizing out the parses rather than just picking the highest
scoring parse. In contrast, the LinearRanker just tries to give a
parse with the correct denotation the highest score, so we should just
pick the denotation of the highest scoring parse in that case.

Finally, the demo had to be updated to reflect the new Ranker interface.